# src/typescript/coreapis-sdk/openapi.yaml
openapi: 3.0.1
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /v1/organizations:
    get:
      tags:
      - Organizations
      summary: List organizations
      operationId: list_organization
      security:
      - SecurityChecker: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 25
          description: Items per page
          default: 25
          title: Limit
        description: Items per page
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Organizations
      summary: Create Organization
      description: Admins can create organizations. Agronomists can create managed
        organizations and are automatically added with agronomist permissions.
      operationId: create_organization
      security:
      - SecurityChecker: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreatePayload'
      responses:
        '200':
          description: Organization Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/organizations/{org_id}:
    get:
      tags:
      - Organizations
      summary: Current Organization Info
      operationId: get_organization
      security:
      - SecurityChecker: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          description: Organization ID
          title: Org Id
        description: Organization ID
      responses:
        '200':
          description: Organization Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/organizations/{org_id}/stats:
    get:
      tags:
      - Organizations
      summary: Get organization statistics
      operationId: get_organization_stats
      security:
      - SecurityChecker: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          description: Organization ID
          title: Org Id
        description: Organization ID
      - name: months
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Rolling time window in months
          default: 3
          title: Months
        description: Rolling time window in months
      responses:
        '200':
          description: Organization statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationStatsResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
      - Organizations
      summary: Update Organization Info
      operationId: update_organization
      security:
      - SecurityChecker: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          description: Organization ID
          title: Org Id
        description: Organization ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdatePayload'
      responses:
        '200':
          description: Organization Info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/organizations/{org_id}/members:
    get:
      tags:
      - Organizations
      summary: List Organization Members
      operationId: list_organization_members
      security:
      - SecurityChecker: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          description: Organization ID
          title: Org Id
        description: Organization ID
      responses:
        '200':
          description: Organization members with roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationMember'
                title: Response List Organization Members
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/organizations/{org_id}/members/{user_id}:
    delete:
      tags:
      - Organizations
      summary: Remove Organization Member
      operationId: remove_organization_member
      security:
      - SecurityChecker: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          description: Organization ID
          title: Org Id
        description: Organization ID
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          description: User ID
          title: User Id
        description: User ID
      responses:
        '200':
          description: Removal status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized to access this resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    AccountTypeEnum:
      type: string
      enum:
      - Admin
      - Agronomist
      - Standard
      title: AccountTypeEnum
...
    Contacts:
      properties:
        email:
          type: string
          title: Email
        phone:
          type: string
          title: Phone
      type: object
      required:
      - email
      - phone
      title: Contacts
...
    ErrorResponse:
      properties:
        status:
          type: integer
          title: Status
        detail:
          anyOf:
          - type: string
          - type: object
          title: Detail
      type: object
      required:
      - status
      - detail
      title: ErrorResponse
...
    FileInfo:
      properties:
        category:
          type: string
          title: Category
        name:
          type: string
          title: Name
      type: object
      required:
      - category
      - name
      title: FileInfo
...
    Organization:
      properties:
        orgId:
          type: string
          title: Orgid
        name:
          type: string
          title: Name
        piva:
          type: string
          title: Piva
        logo:
          type: string
          title: Logo
        cover:
          type: string
          title: Cover
        contacts:
          $ref: '#/components/schemas/Contacts'
        creationTime:
          type: integer
          title: Creationtime
        lastUpdateTime:
          type: integer
          title: Lastupdatetime
      type: object
      required:
      - orgId
      - name
      - piva
      - logo
      - cover
      - contacts
      - creationTime
      - lastUpdateTime
      title: Organization
    OrganizationStatsAgrifield:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        harvest:
          anyOf:
          - type: string
          - type: 'null'
        creationTime:
          anyOf:
          - type: integer
          - type: 'null'
        detectionCount:
          type: integer
          title: Detectioncount
      type: object
      required:
      - id
      - name
      - harvest
      - creationTime
      - detectionCount
      title: OrganizationStatsAgrifield
    OrganizationStatsOrganization:
      properties:
        orgId:
          type: string
          title: Orgid
        name:
          type: string
          title: Name
      type: object
      required:
      - orgId
      - name
      title: OrganizationStatsOrganization
    OrganizationStatsResponse:
      properties:
        organization:
          $ref: '#/components/schemas/OrganizationStatsOrganization'
        window:
          $ref: '#/components/schemas/OrganizationStatsWindow'
        agrifieldCount:
          type: integer
          title: Agrifieldcount
        distinctHarvestCount:
          type: integer
          title: Distinctharvestcount
        distinctHarvests:
          items:
            type: string
          type: array
          title: Distinctharvests
        agrifields:
          items:
            $ref: '#/components/schemas/OrganizationStatsAgrifield'
          type: array
          title: Agrifields
      type: object
      required:
      - organization
      - window
      - agrifieldCount
      - distinctHarvestCount
      - distinctHarvests
      - agrifields
      title: OrganizationStatsResponse
    OrganizationStatsWindow:
      properties:
        months:
          type: integer
          title: Months
        start:
          type: string
          title: Start
        end:
          type: string
          title: End
        startTimestampMs:
          type: integer
          title: Starttimestampms
        endTimestampMs:
          type: integer
          title: Endtimestampms
      type: object
      required:
      - months
      - start
      - end
      - startTimestampMs
      - endTimestampMs
      title: OrganizationStatsWindow
    OrganizationMember:
      properties:
        user:
          $ref: '#/components/schemas/User'
        role:
          type: string
          title: Role
      type: object
      required:
      - user
      - role
      title: OrganizationMember
    OrganizationCreatePayload:
      properties:
        name:
          type: string
          title: Name
        piva:
          type: string
          title: Piva
        contacts:
          $ref: '#/components/schemas/Contacts'
        questionnaire:
          type: object
          additionalProperties: true
          title: Questionnaire
      type: object
      required:
      - name
      - piva
      - contacts
      - questionnaire
      title: OrganizationCreatePayload
    OrganizationUpdatePayload:
      properties:
        name:
          title: Name
          type: string
          nullable: true
        piva:
          title: Piva
          type: string
          nullable: true
        logo:
          $ref: '#/components/schemas/FileInfo'
          nullable: true
        cover:
          $ref: '#/components/schemas/FileInfo'
          nullable: true
        contacts:
          $ref: '#/components/schemas/Contacts'
          nullable: true
      type: object
      title: OrganizationUpdatePayload
    PaginatedResponse:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        limit:
          type: integer
          title: Limit
        data:
          items: {}
          type: array
          title: Data
      type: object
      required:
      - total
      - page
      - limit
      - data
      title: PaginatedResponse
...
    StatusResponse:
      properties:
        status:
          type: integer
          title: Status
        message:
          type: string
          title: Message
      type: object
      required:
      - status
      - message
      title: StatusResponse
    User:
      properties:
        id:
          type: string
          title: Id
        firstName:
          type: string
          title: Firstname
        lastName:
          type: string
          title: Lastname
        email:
          type: string
          title: Email
        emailVerified:
          type: boolean
          title: Emailverified
        enabled:
          type: boolean
          title: Enabled
        accountType:
          $ref: '#/components/schemas/AccountTypeEnum'
        phone:
          type: string
          title: Phone
        piva:
          type: string
          title: Piva
        avatar:
          type: string
          title: Avatar
        organizations:
          items:
            $ref: '#/components/schemas/UserOrgananizationMembership'
          type: array
          title: Organizations
          default: []
        creationTime:
          type: integer
          title: Creationtime
      type: object
      required:
      - id
      - firstName
      - lastName
      - email
      - emailVerified
      - enabled
      - accountType
      - phone
      - piva
      - avatar
      - creationTime
      title: User
...
    UserOrgananizationMembership:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        roles:
          items:
            type: string
          type: array
          title: Roles
      type: object
      required:
      - id
      - name
      - roles
      title: UserOrgananizationMembership
...
  securitySchemes:
    SecurityChecker:
      type: http
      scheme: bearer